move async example from redux-thunk to redux-saga
