+++ examples/async/actions/index.js
@@ -1,9 +1,8 @@
-import fetch from 'isomorphic-fetch'
-
 export const REQUEST_POSTS = 'REQUEST_POSTS'
 export const RECEIVE_POSTS = 'RECEIVE_POSTS'
 export const SELECT_REDDIT = 'SELECT_REDDIT'
 export const INVALIDATE_REDDIT = 'INVALIDATE_REDDIT'
+export const FETCH_POSTS_IF_NEEDED = 'FETCH_POSTS_IF_NEEDED'
 
 export function selectReddit(reddit) {
   return {
@@ -19,14 +18,14 @@ export function invalidateReddit(reddit) {
   }
 }
 
-function requestPosts(reddit) {
+export function requestPosts(reddit) {
   return {
     type: REQUEST_POSTS,
     reddit
   }
 }
 
-function receivePosts(reddit, json) {
+export function receivePosts(reddit, json) {
   return {
     type: RECEIVE_POSTS,
     reddit: reddit,
@@ -35,30 +34,9 @@ function receivePosts(reddit, json) {
   }
 }
 
-function fetchPosts(reddit) {
-  return dispatch => {
-    dispatch(requestPosts(reddit))
-    return fetch(`https://www.reddit.com/r/${reddit}.json`)
-      .then(response => response.json())
-      .then(json => dispatch(receivePosts(reddit, json)))
-  }
-}
-
-function shouldFetchPosts(state, reddit) {
-  const posts = state.postsByReddit[reddit]
-  if (!posts) {
-    return true
-  }
-  if (posts.isFetching) {
-    return false
-  }
-  return posts.didInvalidate
-}
-
 export function fetchPostsIfNeeded(reddit) {
-  return (dispatch, getState) => {
-    if (shouldFetchPosts(getState(), reddit)) {
-      return dispatch(fetchPosts(reddit))
-    }
+  return {
+    type: FETCH_POSTS_IF_NEEDED,
+    reddit
   }
 }
