+++ examples/async/sagas/index.js
@@ -1,0 +1,6 @@
+import { fork } from 'redux-saga/effects'
+import { fetchPostsIfNeededWatcher } from './reddit'
+
+export default function* rootSaga() {
+  yield fork(fetchPostsIfNeededWatcher)
+}
