+++ examples/async/store/configureStore.js
@@ -1,13 +1,18 @@
 import { createStore, applyMiddleware } from 'redux'
-import thunkMiddleware from 'redux-thunk'
+import createSagaMiddleware from 'redux-saga'
 import createLogger from 'redux-logger'
+
 import rootReducer from '../reducers'
+import rootSaga from '../sagas'
 
 export default function configureStore(initialState) {
   const store = createStore(
     rootReducer,
     initialState,
-    applyMiddleware(thunkMiddleware, createLogger())
+    applyMiddleware(
+      createSagaMiddleware(rootSaga),
+      createLogger()
+    )
   )
 
   if (module.hot) {
